Further separation of control base types
Added RotaryDial as a base type
Added LatchButton as a base type

There was some mirrored behaviour in the different definitions of Dials, so it was clear that there was a place for another layer in the inheritance tree.
Added a LatchButton class for the same reasons as above.

NOTE: Accidentally committed an extra static ID for a as-of-yet-non-existent masterBypass control. This will be added in the next commit so I am deeming it not worth reverting this commit to fix one superfluous static int. (Open to discussion on protocols for this if you have a differing opinion!)