Fixing multiple "paint" calls to components: (https://github.com/sblake94/Biquad/issues/1)
Turns out I misunderstood how the "paint" function is called in juce::Components.
I assumed it needed to be called manually, so was calling it from within the ComponentManager.
In actuality it's called from each parent component so in actuality, as long as they are contained within a component that is being painted in the editor, we do not need to call this ourselves.